+++ src/kata/vendingMachine/productManager/IProductManager.java
@@ -1,8 +1,10 @@
 package kata.vendingMachine.productManager;
 
+import java.math.BigDecimal;
+
 import kata.products.enums.ProductType;
 
 public interface IProductManager
 {
-	Double getItemPrice(ProductType productType);
+	BigDecimal getItemPrice(ProductType productType);
 }
